Backport of [JDK-8302191](https://bugs.openjdk.java.net/browse/JDK-8302191). Applies cleanly. It is already committed in trunk.

```
build/linux-x86_64-normal-server-fastdebug/jdk/bin/java -jar ~/azul/jmh-jdk-microbenchmarks-git/micros-jdk8/target/micros-jdk8-1.0-SNAPSHOT.jar org.openjdk.bench.java.lang.FloatDoubleRem -i 2 -r 2 -wi 2 -f 2
unpatched:
Benchmark                            Mode  Cnt    Score    Error  Units
FloatDoubleRem.calcDoubleJava        avgt    4  103.219 ±  4.637  ns/op
FloatDoubleRem.calcFloatJava         avgt    4  106.588 ± 10.357  ns/op
FloatDoubleRem.cornercaseDoubleJava  avgt    4   11.822 ±  4.074  ns/op
FloatDoubleRem.cornercaseFloatJava   avgt    4   11.435 ±  1.355  ns/op
patched:
Benchmark                            Mode  Cnt   Score    Error  Units
FloatDoubleRem.calcDoubleJava        avgt    4  34.712 ± 22.532  ns/op
FloatDoubleRem.calcFloatJava         avgt    4  23.942 ±  5.177  ns/op
FloatDoubleRem.cornercaseDoubleJava  avgt    4   8.080 ±  1.004  ns/op
FloatDoubleRem.cornercaseFloatJava   avgt    4  10.647 ±  0.545  ns/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8302191](https://bugs.openjdk.org/browse/JDK-8302191): Performance degradation for float/double modulo on Linux


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk11u-dev.git pull/1824/head:pull/1824` \
`$ git checkout pull/1824`

Update a local copy of the PR: \
`$ git checkout pull/1824` \
`$ git pull https://git.openjdk.org/jdk11u-dev.git pull/1824/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 1824`

View PR using the GUI difftool: \
`$ git pr show -t 1824`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk11u-dev/pull/1824.diff">https://git.openjdk.org/jdk11u-dev/pull/1824.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk11u-dev/pull/1824#issuecomment-1495942146)