@phohensee 2023-04-06
11u is in maintenance mode, so this backport being purely a performance improvement will be unlikely to be approved by a Maintainer.