@jankratochvil 2023-04-07
This change is related to a GCC fix: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=108922
gcc-4.8 -> gcc-4.9 had a performance regression. This will be fixed in future gcc-13 release. Unaware whether it also gets backported to recent GCC releases.
**This OpenJDK change fixes a performance regression when comparing recent builds of OpenJDK against older OpenJDK releases still be still compiled by <=gcc-4.8.**
This OpenJDK change makes OpenJDK immune against the GCC performance regression. This GCC performance regression is now present in mostly any GCC release in use nowadays.

While my test shows 3x slowdown a customer's testcase shows about 6x slowdown, it depends on the numbers being computed.