@jankratochvil 2023-04-24
`jdk11u-fix-no` has been added at https://bugs.openjdk.org/browse/JDK-8302191